elastic_s3minio/requirements.txt
elasticsearch==7.15.1
PyYAML==6.0.1
Requests==2.31.0
tabulate==0.9.0
urllib3==1.26.16
